request btc part 1
