+++ packages/blockchain-wallet-v4-frontend/src/modals/Btc/RequestBtc/index.js
@@ -1,17 +1,113 @@
 import React from 'react'
-import { compose } from 'redux'
+import { connect } from 'react-redux'
+import { equals, prop } from 'ramda'
+import styled from 'styled-components'
+import { FormattedMessage } from 'react-intl'
+import { bindActionCreators, compose } from 'redux'
 
-import wizardProvider from 'providers/WizardProvider'
+import { actions, model } from 'data'
+import { getData, getInitialValues, getImportedAddresses } from './selectors'
+import Loading from './template.loading'
+import Success from './template.success'
+import DataError from 'components/DataError'
+import { Remote } from 'blockchain-wallet-v4/src'
 import modalEnhancer from 'providers/ModalEnhancer'
-import RequestBtc from './template'
-import FirstStep from './FirstStep'
-import SecondStep from './SecondStep'
+import { Modal, ModalHeader, ModalBody } from 'blockchain-info-components'
 
+const RequestHeader = styled(ModalHeader)`
+  border-bottom: 0;
+  padding-bottom: 0;
+  > div:first-child * {
+    color: ${props => props.theme['brand-primary']};
+  }
+`
+
+const { TRANSACTION_EVENTS } = model.analytics
 class RequestBtcContainer extends React.PureComponent {
-  state = { receiveAddress: '' }
+  state = {
+    requestBuilt: false,
+    showRequestLink: false
+  }
 
   componentDidMount () {
-    this.props.resetStep()
+    this.init()
+  }
+
+  componentDidUpdate (prevProps) {
+    this.props.data.map(x => {
+      if (equals(prop('coin', x), 'PAX')) {
+        this.props.modalActions.closeAllModals()
+        this.props.modalActions.showModal('@MODAL.REQUEST.ETH', {
+          coin: 'PAX',
+          lockboxIndex: this.props.lockboxIndex
+        })
+      } else if (equals(prop('coin', x), 'ETH')) {
+        this.props.modalActions.closeAllModals()
+        this.props.modalActions.showModal('@MODAL.REQUEST.ETH', {
+          coin: 'ETH',
+          lockboxIndex: this.props.lockboxIndex
+        })
+      } else if (equals(prop('coin', x), 'BCH')) {
+        this.props.modalActions.closeAllModals()
+        this.props.modalActions.showModal('@MODAL.REQUEST.BCH', {
+          lockboxIndex: this.props.lockboxIndex
+        })
+      } else if (equals(prop('coin', x), 'XLM')) {
+        this.props.modalActions.closeAllModals()
+        this.props.modalActions.showModal('@MODAL.REQUEST.XLM', {
+          lockboxIndex: this.props.lockboxIndex
+        })
+      }
+    })
+    if (
+      !Remote.Success.is(prevProps.initialValues) &&
+      Remote.Success.is(this.props.initialValues)
+    ) {
+      this.init()
+    }
+  }
+
+  init = () => {
+    this.props.initialValues.map(x => {
+      this.props.formActions.initialize('requestBtc', x)
+    })
+  }
+
+  handleOpenLockbox = () => {
+    this.props.requestBtcActions.openLockboxAppClicked()
+  }
+
+  onToggleMakeRequestLink = () => {
+    this.setState({
+      requestBuilt: false,
+      showRequestLink: !this.state.showRequestLink
+    })
+  }
+
+  handleSubmit = e => {
+    e.preventDefault()
+    const {
+      accountIdx,
+      addressIdx,
+      message,
+      receiveAddress
+    } = this.props.data.getOrElse({})
+    this.props.requestBtcActions.firstStepSubmitClicked({
+      accountIdx,
+      addressIdx,
+      message
+    })
+    this.setState({
+      receiveAddress
+    })
+    this.props.analyticsActions.logEvent([...TRANSACTION_EVENTS.REQUEST, 'BTC'])
+  }
+
+  handleRefresh = () => {
+    const { btcDataActions, initialValues } = this.props
+    if (!Remote.Success.is(initialValues)) return btcDataActions.fetchData()
+
+    this.init()
   }
 
   setReceiveAddress = addr => {
@@ -19,23 +115,67 @@ class RequestBtcContainer extends React.PureComponent {
   }
 
   render () {
-    const { receiveAddress } = this.state
-    const { step, position, total, closeAll, ...rest } = this.props
+    const { data, importedAddresses, position, total, closeAll } = this.props
+
+    const content = data.cata({
+      Success: value => (
+        <Success
+          type={value.type}
+          message={value.message}
+          addressIdx={value.addressIdx}
+          accountIdx={value.accountIdx}
+          receiveAddress={value.receiveAddress}
+          handleOpenLockbox={this.handleOpenLockbox}
+          onToggleMakeRequestLink={this.onToggleMakeRequestLink}
+          handleSubmit={this.handleSubmit}
+          importedAddresses={importedAddresses}
+          excludeLockbox={value.excludeLockbox}
+          setReceiveAddress={this.setReceiveAddress}
+          showRequestLinkBuilder={this.state.showRequestLink}
+          closeAll={closeAll}
+        />
+      ),
+      NotAsked: () => <DataError onClick={this.handleRefresh} />,
+      Failure: () => <DataError onClick={this.handleRefresh} />,
+      Loading: () => <Loading />
+    })
 
     return (
-      <RequestBtc position={position} total={total} closeAll={closeAll}>
-        {step === 1 && (
-          <FirstStep setReceiveAddress={this.setReceiveAddress} {...rest} />
-        )}
-        {step === 2 && <SecondStep receiveAddress={receiveAddress} {...rest} />}
-      </RequestBtc>
+      <Modal size='small' position={position} total={total}>
+        <RequestHeader icon='request' onClose={closeAll}>
+          <FormattedMessage
+            id='modals.requestbtc.title'
+            defaultMessage='Request Bitcoin'
+          />
+        </RequestHeader>
+        <ModalBody>{content}</ModalBody>
+      </Modal>
     )
   }
 }
+const mapStateToProps = (state, ownProps) => ({
+  initialValues: getInitialValues(state, ownProps),
+  data: getData(state),
+  importedAddresses: getImportedAddresses(state)
+})
+
+const mapDispatchToProps = dispatch => ({
+  analyticsActions: bindActionCreators(actions.analytics, dispatch),
+  requestBtcActions: bindActionCreators(
+    actions.components.requestBtc,
+    dispatch
+  ),
+  btcDataActions: bindActionCreators(actions.core.data.btc, dispatch),
+  modalActions: bindActionCreators(actions.modals, dispatch),
+  formActions: bindActionCreators(actions.form, dispatch)
+})
 
 const enhance = compose(
   modalEnhancer('@MODAL.REQUEST.BTC'),
-  wizardProvider('requestBtc', 2)
+  connect(
+    mapStateToProps,
+    mapDispatchToProps
+  )
 )
 
 export default enhance(RequestBtcContainer)
