+++ packages/blockchain-wallet-v4-frontend/src/modals/Btc/RequestBtc/template.loading.js
