+++ packages/blockchain-wallet-v4-frontend/src/modals/Btc/RequestBtc/template.shareLink.js
@@ -1,0 +1,134 @@
+import React, { Fragment } from 'react'
+import PropTypes from 'prop-types'
+import styled from 'styled-components'
+import { FormattedMessage } from 'react-intl'
+import { reduxForm } from 'redux-form'
+import CopyToClipBoard from 'react-copy-to-clipboard'
+
+import { Button, Link, Text } from 'blockchain-info-components'
+import { Form } from 'components/Form'
+import CoinDisplay from 'components/Display/CoinDisplay'
+import FiatDisplay from 'components/Display/FiatDisplay'
+
+const Container = styled.div`
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  padding: 20px 0;
+  margin-top: 5px;
+  margin-bottom: 20px;
+  background-color: ${props => props.theme['white-blue']};
+
+  & > * {
+    padding: 5px 0;
+  }
+`
+const LinkContainer = styled.div`
+  width: 100%;
+  margin-top: 20px;
+  box-sizing: border-box;
+  padding: 20px 15px 0px 15px;
+  text-align: center;
+  word-wrap: break-word;
+  border-top: 1px solid ${props => props.theme['gray-1']};
+`
+const SubHeader = styled(Text)`
+  margin-bottom: 20px;
+`
+const Footer = styled.div`
+  margin-top: 20px;
+`
+
+const SecondStep = props => {
+  const { previousStep, ...rest } = props
+  const { handleSubmit, satoshis, message, link, active } = rest
+
+  return (
+    <Fragment>
+      <Form onSubmit={handleSubmit}>
+        <SubHeader size='14px' weight={400}>
+          <FormattedMessage
+            id='modals.requestbitcoin.secondstep.explain'
+            defaultMessage='Send the link below to your friend or contact and they will be able to send bitcoin directly to your wallet.'
+          />
+        </SubHeader>
+        <Text size='14px' weight={500}>
+          <FormattedMessage
+            id='modals.requestbitcoin.secondstep.payment'
+            defaultMessage='Payment Request:'
+          />
+        </Text>
+        <Container>
+          <CoinDisplay size='28px' weight={500} color='received' coin='BTC'>
+            {satoshis}
+          </CoinDisplay>
+          <FiatDisplay size='20px' weight={400} color='received' coin='BTC'>
+            {satoshis}
+          </FiatDisplay>
+          <Text size='16px'>
+            <FormattedMessage
+              id='modals.requestbitcoin.secondstep.message'
+              defaultMessage='For "{message}"'
+              values={{ message: message }}
+            />
+          </Text>
+          <LinkContainer>
+            <Text size='12px' weight={400} data-e2e='requestBtcLinkText'>
+              {link}
+            </Text>
+          </LinkContainer>
+        </Container>
+        <CopyToClipBoard text={link}>
+          <Button
+            type='submit'
+            nature={active ? 'copy' : 'primary'}
+            fullwidth
+            data-e2e='requestBtcCopyLinkButton'
+          >
+            {active ? (
+              <FormattedMessage
+                id='modals.requestbitcoin.secondstep.copied'
+                defaultMessage='Copied!'
+              />
+            ) : (
+              <FormattedMessage
+                id='modals.requestbitcoin.secondstep.copy'
+                defaultMessage='Copy Link'
+              />
+            )}
+          </Button>
+        </CopyToClipBoard>
+      </Form>
+      <Footer>
+        <Link
+          onClick={previousStep}
+          size='13px'
+          weight={400}
+          data-e2e='requestBtcGoBack'
+        >
+          <FormattedMessage
+            id='scenes.requestbitcoin.secondstep.back'
+            defaultMessage='Go Back'
+          />
+        </Link>
+      </Footer>
+    </Fragment>
+  )
+}
+
+SecondStep.propTypes = {
+  position: PropTypes.number.isRequired,
+  total: PropTypes.number.isRequired,
+  closeAll: PropTypes.func.isRequired,
+  active: PropTypes.bool.isRequired,
+  satoshis: PropTypes.string.isRequired,
+  link: PropTypes.string.isRequired,
+  message: PropTypes.string.isRequired,
+  previousStep: PropTypes.func.isRequired,
+  onSubmit: PropTypes.func.isRequired
+}
+
+export default reduxForm({ form: 'requestBtc', destroyOnUnmount: false })(
+  SecondStep
+)
